Determine principal name on app upgrades from 0.5
The username that is stored in the account settings is not necessarily
the principal name which we require to build the principal URL.
So query server for the principal name (this has to happen async).

We can't use the DavResourceFinder here, as bound services are not
allowed in BroadCastReceivers.
That's also why we need to change HttpClient to not use a custom cert
manager. Disabling the cert manager during the upgrade should be ok, as
the old app did not allow custom certificates.

If we fail to determine the principal URL, delete the old account and
continue without migrating the account settings.
Leaving the old accounts around would break the app.